backend/auth_gsheets.py
"""
Authentication module for Google Sheets backend.
Modified to work with dictionary users instead of SQLAlchemy models.
"""
from datetime import datetime, timedelta
import bcrypt
from jose import JWTError, jwt
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from config import settings
from crud_gsheets import GSheetsCRUD

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")

def hash_password(password: str) -> str:
    return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()

def verify_password(plain: str, hashed: str) -> bool:
    return bcrypt.checkpw(plain.encode(), hashed.encode())

def create_access_token(data: dict, expires_delta: timedelta | None = None) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)

def get_crud():
    """Dependency to get CRUD instance."""
    return GSheetsCRUD()

async def get_current_user(
    token: str = Depends(oauth2_scheme), 
    crud: GSheetsCRUD = Depends(get_crud)
) -> dict:
    """Get current user from JWT token using Google Sheets CRUD."""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
        email: str = payload.get("sub")
        if email is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    
    user = crud.get_user_by_email(email)
    if user is None:
        raise credentials_exception
    return user

def require_role(*roles: str):
    """Decorator to require specific roles."""
    def role_checker(current_user: dict = Depends(get_current_user)):
        if current_user["role"].value not in roles:
            raise HTTPException(status_code=403, detail="Not enough permissions")
        return current_user
    return role_checker