backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
sqlalchemy==2.0.49
alembic==1.18.4
asyncpg==0.31.0
aiosqlite==0.22.1
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
aiosmtplib==5.1.0
jinja2==3.1.6
apscheduler==3.11.2
google-api-python-client==2.173.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.2.0
oauth2client==4.1.3
